Adopt standalone utils functions into Runner
This change moves a chain of standalone functions into Runner in order
to make it possible to access app instance from it.

_emit_matches --> find_children --> play_children --> _taskshandlers_children --> _roles_children --> _rolepath
